[AND-172] Add signature info to package info extensions results
Also break separate extension functions by class name to organise them
better.
